add logging for no files

+++ drop2apex.py
@@ -32,12 +32,16 @@ pwd = cfg[cfg_section]["pwd"]
 sftph = sftp_handler.SftpHandler(host, user, pwd)
 sftph.set_dir(from_dir)
 files = sftph.listdir()
-for file in files:
-    sftph.read_file(file, workdir=to_dir)
-    sftph.remove_file(file)
-    logline = "{script};{date};{time}FROM;databestand {file} van de dropserver afgehaald\n".format(script=script,
-                                                                                                   date=ds, time=ts,
-                                                                                                   file=file)
+if len(files) > 0:
+    for file in files:
+        sftph.read_file(file, workdir=to_dir)
+        sftph.remove_file(file)
+        logline = "{script};{date};{time}FROM;databestand {file} van de dropserver afgehaald\n".format(script=script,
+                                                                                                       date=ds, time=ts,
+                                                                                                       file=file)
+        logfh.write(logline)
+else:
+    logline = "{script};{date};{time}FROM;geen databestanden aanwezig.\n".format(script=script, date=ds, time=ts)
     logfh.write(logline)
 sftph.close_connection()
 logfh.close()
